backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
redis==5.0.1
python-multipart==0.0.6
Pillow==10.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.6.1
motor==3.3.2
aiofiles==23.2.1
httpx==0.26.0
solders==0.19.1
anchorpy==0.18.0 